requirements.txt
flask
flask-cors
psycopg2-binary
gunicorn
google-generativeai
python-dotenv
requests